+++ src/main/java/pl/project/plannerapp/service/DietService.java
@@ -1,6 +1,5 @@
 package pl.project.plannerapp.service;
 
-import org.hibernate.validator.constraints.NotEmpty;
 import pl.project.plannerapp.DTO.DietDTO;
 
 import java.util.List;
@@ -8,7 +7,9 @@ import java.util.Optional;
 
 public interface DietService {
     List<DietDTO> getAll();
-    void put(int id, DietDTO dietDTO);
-    void delete(int id);
-    Optional<DietDTO> getById(int id);
+    void put(Long id, DietDTO dietDTO);
+
+    void delete(Long id);
+
+    Optional<DietDTO> getById(Long id);
 }
